Add DialogService
Create a interface that allows viewmodels to show dialogs.

Create interface.
Implement interface in folder IT4ClubCar. It uses the method DisplayAlert. Perhaps in the future it will use the native warning methods of each platform. Called IT4ClubCarDialogService.
In the BaseViewModel create a property of type IDialogService so it can be used by all viewmodels.
In every viewmodel created ask for a IDialogService in the constructor.
In the App.cs, using Unity, inject a IT4ClubCarDialogService when some class asks for a variable of type IDialogService.