Bug Fix - Exception throw when using the DialogService
An exception is throw when the PaginaInicialViewModel uses the DialogService. This viewmodel is used by the first page to appear. The dialog service uses the main page to display alerts. By the fact that when the PaginaInicialView is being created, the MainPage isn't defined yet, an exception is throw because the MainPage is null when the PaginaInicialViewModel tries to use it.

Quick fix :
In the dialog service check if MainPage isn't null.

Problem:
The message that was asked to show when the MainPage is null, will never be show.